Moved header and private route to component folder, created initial styles
