+++ gamersnet_frontend/src/components/header/styles.css
@@ -1,0 +1,19 @@
+.header {
+    font-size: 1.25rem;
+    padding: 1rem 2rem 1rem 2rem;
+    border-bottom: 1px solid #d4d4d4;
+    margin-bottom: 2rem;
+}
+
+.header > * {
+    display: inline;
+}
+
+.header a {
+    text-decoration: none;
+    color: rgb(53, 53, 53);
+}
+
+.header .header-link {
+    float: right;
+}
