+++ gamersnet_frontend/src/cookieCheck.js
@@ -1,0 +1,9 @@
+export default function cookieCheck() {
+    for (let cookie of document.cookie.split('; ')) {
+        if (cookie.split('=')[0] === 'token') {
+            return true;
+        }
+    }
+
+    return false;
+}
