+++ gamersnet_frontend/src/index.js
@@ -1,25 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {Link, Route, BrowserRouter} from 'react-router-dom';
+import {Route, BrowserRouter} from 'react-router-dom';
 
-import PrivateRoute from './PrivateRoute';
+import './styles.css';
+
+import PrivateRoute from './components/PrivateRoute';
 
 // import all the pages from their component folders
 import Home from './components/home';
 import SignIn from './components/signin';
 import Password from './components/password';
 
-// function to render the "header" (just the links at the top)
-// this returns a set of <a> HTML tags that tell react to go to a page (using react-router)
-function Header() {
-	return (
-		<div>
-			<Link to = '/'>home</Link><br />
-			<Link to = '/signin'>sign in</Link><br />
-			<Link to = '/password'>password</Link><br />
-		</div>
-	);
-}
+// import header
+import Header from './components/header';
 
 // function that defines what to do when user goes to a specific page/
 // this function binds the /path to the component it belongs to
