+++ gamersnet_frontend/src/styles.css
@@ -1,0 +1,10 @@
+* {
+    margin: 0;
+    padding: 0;
+}
+
+body {
+    font-family: Arial;
+    font-size: 18px;
+    background-color: #f5f5f5;
+}
